rag-jobs-free/requirements.txt
langchain>=0.2.12
langchain-community>=0.2.10
chromadb>=0.5.5
tiktoken>=0.7.0
python-dotenv>=1.0.1
pandas>=2.2.2
rich>=13.7.1
sentence-transformers>=3.0.1
langchain-ollama>=0.1.0
groq>=0.9.0
